pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codestory"
version = "0.1.0"
description = "An AI-powered abstraction layer above Git. Focus on vibecoding, let codestory handle the rest."
authors = [
  { name = "Adem Can", email = "ademfcan@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "typer-slim>=0.12.0",
  "colorama>=0.4.6",
  "loguru>=0.7.3",
  "tree-sitter>=0.21.0",
  "tree-sitter-language-pack>=0.1.0",
  "platformdirs>=4.5.0",
  "httpx>=0.28.1",
  "tqdm>=4.67.1"
]

[project.scripts]
cst = "codestory.cli:run_app"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "mypy",
  "mkdocs",
  "mkdocs-material",
  "mkdocs-typer2"
]

[tool.setuptools.package-data]
codestory = ["resources/*"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "tree_sitter.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py310"
src = ["src"]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "I001", # Enforce absolute imports
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "B904",  # force raise from ...
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]


[tool.ruff.lint.isort]
known-local-folder = ["codestory"]

[tool.pytest.ini_options]
pythonpath = ["src"]